+++ sszgen/generator/marshal.go
@@ -132,7 +132,7 @@ func (v *Value) marshalVector() (str string) {
 	obj := v.typ.(*Vector)
 	inner.name = fmt.Sprintf("%s[ii]", v.name)
 
-	tmpl := `{{.validate}}for ii := 0; ii < {{.size}}; ii++ {
+	tmpl := `{{.validate}}for ii := uint64(0); ii < {{.size}}; ii++ {
 		{{.marshal}}
 	}`
 	return execTmpl(tmpl, map[string]interface{}{
